requirements.txt
# ============================================
# Gmail AI Assistant - Requirements
# ============================================

# Core Dependencies
# ============================================
python-dotenv==1.0.1
streamlit==1.29.0

# Google API & Authentication
# ============================================
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2

# LangChain Framework
# ============================================
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-google-genai==0.0.6

# Embeddings & Vector Store
# ============================================
sentence-transformers==2.2.2
faiss-cpu==1.8.0
transformers==4.36.0
torch==2.1.0

# HTML Parsing & Text Processing
# ============================================
beautifulsoup4==4.12.2
lxml==5.1.0
html5lib==1.1

# Additional Utilities
# ============================================
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
urllib3==2.1.0

# Optional: For better performance
# ============================================
# faiss-gpu==1.8.0  # Uncomment for GPU acceleration (requires CUDA)
# protobuf==4.25.1  # May be needed for some configurations